requirements.txt
gradio==4.13.0
torch==2.1.1
torchaudio==2.1.1
faster-whisper==1.0.1
tts>=0.22.0
langid
noisereduce
pedalboard
pydub
ffmpeg-python
soundfile
cutlet
fugashi[unidic-lite]
loguru
omegaconf==2.3.0
resampy==0.4.2
tabulate==0.8.10
requests
faiss-cpu
pyworld
torchcrepe
praat-parselmouth>=0.4.2
translators
transformers==4.36.2
deepl
requests
pysubs2
fastapi
srt
moviepy
ffmpeg-python==0.2.0
#onnx
#onnxruntime-gpu
